AI/trashdig/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/trashdig"]

[project]
name = "trashdig"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-adk",
    "pathspec>=1.0.4",
    "semgrep>=1.159.0",
    "textual>=8.2.3",
    "textual-autocomplete>=4.0.6",
    "tomli-w>=1.2.0",
    "tree-sitter>=0.25.2",
    "tree-sitter-c-sharp>=0.23.1",
    "tree-sitter-go>=0.25.0",
    "tree-sitter-javascript>=0.25.0",
    "tree-sitter-python>=0.25.0",
]

[project.scripts]
trashdig = "trashdig.main:main"

[tool.pytest.ini_options]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "mypy>=1.20.1",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.10",
]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple/"
default = true

[tool.ruff]

[tool.ruff.lint]
# Select default rules plus isort (I)
select = ["E4", "E7", "E9", "F", "I"]

[tool.ruff.lint.isort]
known-first-party = ["trashdig"]
